+++ tests/units/test_health_endpoint.py
@@ -11,14 +11,24 @@ from reflex.model import get_db_status
 from reflex.utils.prerequisites import get_redis_status
 
 
+def _get_async_function(func):
+    async def _async_func(*args, **kwargs):
+        return func(*args, **kwargs)
+
+    return _async_func
+
+
 @pytest.mark.asyncio
 @pytest.mark.parametrize(
     ("mock_redis_client", "expected_status"),
     [
         # Case 1: Redis client is available and responds to ping
-        (Mock(ping=lambda: None), {"redis": True}),
+        (Mock(ping=_get_async_function(lambda: None)), {"redis": True}),
         # Case 2: Redis client raises RedisError
-        (Mock(ping=lambda: (_ for _ in ()).throw(RedisError)), {"redis": False}),
+        (
+            Mock(ping=_get_async_function(lambda: (_ for _ in ()).throw(RedisError))),
+            {"redis": False},
+        ),
         # Case 3: Redis client is not used
         (None, {"redis": None}),
     ],
@@ -26,9 +36,9 @@ from reflex.utils.prerequisites import get_redis_status
 async def test_get_redis_status(
     mock_redis_client, expected_status, mocker: MockerFixture
 ):
-    # Mock the `get_redis_sync` function to return the mock Redis client
-    mock_get_redis_sync = mocker.patch(
-        "reflex.utils.prerequisites.get_redis_sync", return_value=mock_redis_client
+    # Mock the `get_redis` function to return the mock Redis client
+    mock_get_redis = mocker.patch(
+        "reflex.utils.prerequisites.get_redis", return_value=mock_redis_client
     )
 
     # Call the function
@@ -36,7 +46,7 @@ async def test_get_redis_status(
 
     # Verify the result
     assert status == expected_status
-    mock_get_redis_sync.assert_called_once()
+    mock_get_redis.assert_called_once()
 
 
 @pytest.mark.asyncio
